# src/main/resources/swagger_specification.yaml
openapi: 3.0.0
info:
  version: 0.0.2
  title: Superheroes
  description: A simple Superheroes API
paths:
  /superheroes:
    post:
      description: Creates a new superhero in the store. Duplicates are allowed
      operationId: addSuperhero
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/newSuperhero"
        description: Superhero to add to the store
        required: true
      responses:
        "201":
          description: Superhero response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/superhero"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorModel"
  "/superheroes/{id}":
    get:
      description: Returns a superhero
      operationId: findSuperheroById
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Superhero response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/superhero"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorModel"
servers:
  - url: /
components:
  parameters:
    id:
      name: id
      description: ID of superhero to fetch
      in: path
      required: true
      schema:
        type: integer
        format: int64
  schemas:
    superhero:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        tag:
          type: string
    newSuperhero:
      type: object
      required:
        - name
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        tag:
          type: string
    errorModel:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string